final/requirements.txt
kafka-python @ git+https://github.com/dpkp/kafka-python.git@a6d0579d3cadd3826dd364b01bc12a2173139abc
psycopg2-binary
requests
sqlalchemy
pyspark
numpy